+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add validation API grape class
 * Add call to get last_synced_at for different sync routines
 * Allow routing in completed_dispersal endpoint
+* Allow collecting invoices for collected orders
 
 ### 2.0.0